requirements.txt
# MemDoc Dependencies

# Web Framework
Flask==3.0.0
Eel==0.16.0

# Markdown Processing
PyYAML==6.0.1
markdown2==2.4.12

# PDF Generation
WeasyPrint==60.2

# Image Processing
Pillow==10.1.0

# Development (optional)
# pytest==7.4.3
